quirc: do not create no-version .so symlink if it exists already
Missed that a bump in #328420 changes this part of installation

+++ pkgs/tools/graphics/quirc/default.nix
@@ -55,7 +55,7 @@ stdenv.mkDerivation (finalAttrs: {
     dylib=$out/lib/libquirc.${finalAttrs.version}.dylib
     ${stdenv.cc.targetPrefix}install_name_tool -id "$dylib" "$dylib"
   '' else ''
-    ln -s $out/lib/libquirc.so.* $out/lib/libquirc.so
+    test -e "$out/lib/libquirc.so" || ln -s $out/lib/libquirc.so.* $out/lib/libquirc.so
   '');
 
   meta = {
